mudança de botões
