+++ src/botoes/botoes_main.py
@@ -32,6 +32,16 @@ class Widget:
     def __repr__(self):
         return f"Botão '{self.texto}'"
 
+class BotaoRecomecar(Widget):
+    def __init__(self, x, y, width, height, texto, func):
+        super().__init__(x, y, width, height, texto)
+        self.on_click = func
+
+class BotaoMenu(Widget):
+    def __init__(self, x, y, width, height, texto, func):
+        super().__init__(x, y, width, height, texto)
+        self.on_click = func
+
 class BotaoComprar(Widget):
     def __init__(self, x, y, width, height, texto, func):
         super().__init__(x, y, width, height, texto)
